+++ trixnity-messenger-compose-view/src/androidMain/kotlin/de/connect2x/trixnity/messenger/compose/view/form/rememberHiddenRegistrationForm.android.kt
@@ -1,0 +1,8 @@
+package de.connect2x.trixnity.messenger.compose.view.form
+
+import androidx.compose.runtime.Composable
+
+@Composable
+internal actual fun rememberHiddenRegistrationForm(uniqueId: String?): HiddenRegistrationForm {
+    return NoopHiddenRegistrationForm
+}