+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/form/HiddenRegistrationForm.kt
@@ -1,0 +1,9 @@
+package de.connect2x.trixnity.messenger.compose.view.form
+
+internal interface HiddenRegistrationForm {
+    fun submit(username: String, password: String)
+}
+
+internal data object NoopHiddenRegistrationForm : HiddenRegistrationForm {
+    override fun submit(username: String, password: String) {}
+}