+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/form/rememberHiddenRegistrationForm.kt
@@ -1,0 +1,5 @@
+package de.connect2x.trixnity.messenger.compose.view.form
+
+import androidx.compose.runtime.Composable
+
+@Composable internal expect fun rememberHiddenRegistrationForm(uniqueId: String? = null): HiddenRegistrationForm