+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SDK: Added check for downloadsDisabled before downloading file
 - SDK: granular download limits for syncing media files in `MatrixMessengerConfiguration`
 - SDK,UI,WEB: Experimental voice messages
+- UI: support password managers on Web
 
 ### Changed